refactor(db): remove @prisma/client usage in favor of @taiyomoe/db exports
