+++ apps/io-worker/src/handlers/create-media.handler.ts
@@ -1,4 +1,4 @@
-// import type { TaskStatus } from "@prisma/client"
+// import type { TaskStatus } from "@taiyomoe/db"
 // import { db } from "@taiyomoe/db"
 // import {
 //   BaseCoversService,
