+++ apps/io-worker/src/services/md.worker-service.ts
@@ -1,4 +1,4 @@
-import type { Prisma } from "@prisma/client"
+import type { Prisma } from "@taiyomoe/db"
 import type { Manga } from "mangadex-full-api"
 import { MdUtils } from "~/utils/md.utils"
