+++ apps/io-worker/src/utils/update-task-status.ts
@@ -1,4 +1,4 @@
-import type { TaskStatus } from "@prisma/client"
+import type { TaskStatus } from "@taiyomoe/db"
 import { db } from "@taiyomoe/db"
 import { rawQueue } from "@taiyomoe/messaging"
 import type { ImportChapterMessageInput } from "@taiyomoe/types"
