+++ apps/web/package.json
@@ -13,7 +13,6 @@
   "dependencies": {
     "@hookform/resolvers": "^5.2.2",
     "@marsidev/react-turnstile": "^1.3.0",
-    "@prisma/client": "^6.16.1",
     "@radix-ui/react-dialog": "^1.1.15",
     "@radix-ui/react-dropdown-menu": "^2.1.16",
     "@radix-ui/react-hover-card": "^1.1.15",
@@ -24,6 +23,7 @@
     "@tailwindcss/postcss": "^4.1.13",
     "@taiyomoe/auth": "workspace:^",
     "@taiyomoe/config": "workspace:^",
+    "@taiyomoe/db": "workspace:^",
     "@taiyomoe/schemas": "workspace:^",
     "@taiyomoe/trpc": "workspace:^",
     "@taiyomoe/ui": "workspace:^",
