+++ packages/auth/src/config.ts
@@ -1,6 +1,6 @@
-import type { Roles } from "@prisma/client"
 import { cacheClient } from "@taiyomoe/cache"
 import { db } from "@taiyomoe/db"
+import type { Roles } from "@taiyomoe/db/browser"
 import {
   USERNAME_MAX_LENGTH,
   USERNAME_MIN_LENGTH,
