+++ packages/meilisearch/README.md
@@ -1,0 +1,11 @@
+# `@taiyomoe/meilisearch`
+
+This package contains the Meilisearch integration for the project.
+
+## Seed
+
+```ts
+import { meilisearchClient } from "@taiyomoe/meilisearch"
+
+await meilisearchClient.seed()
+```
